Fix Framer custom login
