+++ apps/web/app/(ee)/partners.dub.co/(auth)/[programSlug]/login/page.tsx
@@ -1,6 +1,6 @@
 export {
   generateMetadata,
   generateStaticParams,
-} from "../../../../(apply)/[programSlug]/layout";
+} from "../../../(apply)/[programSlug]/layout";
 
 export { default } from "../../login/page";
