+++ apps/web/app/(ee)/partners.dub.co/(auth)/auth/reset-password/[token]/page.tsx
